requirements.txt
langchain==0.1.0
python-dotenv==1.0.0
langchain-community==0.0.10
chromadb==0.4.22
fastapi==0.115.11
uvicorn==0.34.0
jinja2==3.1.3
python-multipart==0.0.9
alpha_vantage==2.3.1
pandas==2.2.1 